+++ src/app/(dashboard)/dashboard/create/page.tsx
@@ -9,6 +9,17 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
 import { useRouter } from 'next/navigation';
+import { format } from "date-fns"
+import { Calendar as CalendarIcon } from "lucide-react"
+
+import { cn } from "@/lib/utils"
+import { Calendar } from "@/components/ui/calendar"
+import {
+    Popover,
+    PopoverContent,
+    PopoverTrigger,
+} from "@/components/ui/popover"
+import { Switch } from '@/components/ui/switch';
 
 const CreateNewForm = () => {
 
@@ -18,7 +29,10 @@ const CreateNewForm = () => {
     const [originalUrl, setOriginalUrl] = useState('');
     const [title, setTitle] = useState('');
     const [customBackHalf, setCustomBackHalf] = useState('');
+    const [customExpiry, setCustomExpiry] = useState<Date | undefined>(undefined);
     const [loading, setLoading] = useState(false);
+    const [passwordProtected, setPasswordProtected] = useState(false);
+    const [password, setPassword] = useState('');
 
     const domain = process.env.NEXT_PUBLIC_SHORT_DOMAIN;
 
@@ -31,8 +45,13 @@ const CreateNewForm = () => {
         const data = {
             originalUrl,
             title,
-            urlId: customBackHalf || undefined // Optional custom back-half
+            urlId: customBackHalf || undefined, // Optional custom back-half
+            urlExpiry: customExpiry || undefined, // Optional custom expiry
+            isPasswordProtected: passwordProtected,
+            password: passwordProtected ? password : undefined,
         };
+
+        console.log(data);
 
         try {
             setLoading(true);
@@ -43,11 +62,11 @@ const CreateNewForm = () => {
                 setOriginalUrl(''); // Reset form fields
                 setTitle('');
                 setCustomBackHalf('');
-                router.replace('/dashboard/urls/')
+                router.replace(`/dashboard/urls/${response.data.urlId}`)
             } else {
                 toast({ title: "Error", description: response.data.message || "Something went wrong", variant: "destructive" });
             }
-        } catch (error:any) {
+        } catch (error: any) {
             console.log(error);
             toast({ title: "Error", description: error.response.data.message || "Failed to create short URL.", variant: "destructive" });
         } finally {
@@ -109,6 +128,70 @@ const CreateNewForm = () => {
                                         <p className="text-sm text-gray-500 mt-1">You can create custom back-halves.</p>
                                     </div>
                                 </div>
+                                <div className="space-y-4 mt-4">
+                                    <div>
+                                        <label className="block mb-2 font-medium">Link Expiry</label>
+                                        <div className="flex space-x-2">
+                                            <Popover>
+                                                <PopoverTrigger asChild>
+                                                    <Button
+                                                        variant={"outline"}
+                                                        className={cn(
+                                                            "w-[280px] justify-start text-left font-normal",
+                                                            !customExpiry && "text-muted-foreground"
+                                                        )}
+                                                    >
+                                                        <CalendarIcon className="mr-2 h-4 w-4" />
+                                                        {customExpiry ? format(customExpiry, "PPP") : <span>Pick a Expiry date</span>}
+                                                    </Button>
+                                                </PopoverTrigger>
+                                                <PopoverContent className="w-auto p-0">
+                                                    <Calendar
+                                                        mode="single"
+                                                        selected={customExpiry}
+                                                        onSelect={setCustomExpiry}
+                                                        disabled={(date) => {
+                                                            const today = new Date();
+                                                            return date <= today;
+                                                        }}
+                                                        initialFocus
+                                                    />
+                                                </PopoverContent>
+                                            </Popover>
+                                        </div>
+                                        <p className="text-sm text-gray-500 mt-1">Default is 100 years.</p>
+                                    </div>
+                                </div>
+                                <div className='mt-4'>
+                                    <h2 className="text-xl font-semibold mb-4">Password Protection</h2>
+                                    <div className="space-y-4 mt-4">
+                                        <div>
+                                            <div className="flex items-center space-x-2">
+                                                <label htmlFor="passwordToggle" className="toggle-label">Add password</label>
+                                                <Switch
+                                                    id="passwordToggle"
+                                                    checked={passwordProtected}
+                                                    onCheckedChange={setPasswordProtected}
+                                                />
+                                            </div>
+                                        </div>
+                                        {passwordProtected && (
+                                            <div className="mt-4">
+                                                <label className="block mb-2 font-medium">Password</label>
+                                                <Input
+                                                    type="password"
+                                                    placeholder="Enter password"
+                                                    value={password}
+                                                    onChange={(e) => setPassword(e.target.value)}
+                                                    minLength={5}
+                                                    required
+                                                />
+                                                <p className="text-sm text-gray-500 mt-1">Password cannot be empty.</p>
+                                            </div>
+                                        )}
+                                    </div>
+                                </div>
+
                             </div>
                         </div>
                     </div>